config: set correct bed dimensions based on the glass position
As the motor is on right side, the X maximum is based on the actual
bed size. That way the X0 comes correctly on the glass. The actual
printing area is kept limited to the glass with the help of slicer.

When the bed reaches the Y endstop, nozzle is not on the glass (it is
even not on the bed, it is in the air). Setting the Y_MIN_POS to a
correct negative value, so that Y0 comes correctly on the glass.

https://www.reprap.org/wiki/Configuring_Marlin_Bed_Dimensions

Signed-off-by: Bhavin Gandhi <bhavin7392@gmail.com>